+++ executables/GLFW/TestGLFW_TextureArrayAnimation/src/main.cpp
@@ -175,7 +175,7 @@ protected:
 		hints.decorated = 1;
 	}
 
-	virtual bool Tick(double delta_time) override
+	virtual bool Tick(float delta_time) override
 	{
 		time += delta_time;
 		fps_view_controller.Tick(glfw_window, delta_time);
