+++ executables/LUDUMDARE/Ludum43/src/Ludum43Player.h
@@ -35,17 +35,17 @@ protected:
 	/** override */
 	virtual void OnLevelChanged(death::GameLevel * new_level, death::GameLevel * old_level, death::GameLevelInstance * new_level_instance) override;
 	/** override */
-	virtual void TickPlayerDisplacement(double delta_time) override;
+	virtual void TickPlayerDisplacement(float delta_time) override;
 	/** override */
-	virtual void InternalHandleGamepadInputs(double delta_time, chaos::MyGLFW::GamepadData const * gpd) override;
+	virtual void InternalHandleGamepadInputs(float delta_time, chaos::MyGLFW::GamepadData const * gpd) override;
 	/** override */
-	virtual void HandleKeyboardInputs(double delta_time) override;
+	virtual void HandleKeyboardInputs(float delta_time) override;
 
 	/** update player displacement with the stick values that are stored */
-	virtual void UpdatePlayerAcceleration(double delta_time);
+	virtual void UpdatePlayerAcceleration(float delta_time);
 	/** cooldown the weapon */
-	void TickCooldown(double delta_time);
-	void TickDashValues(double delta_time);
+	void TickCooldown(float delta_time);
+	void TickDashValues(float delta_time);
 
 	/** change dash mode */
 	void SetDashMode(bool dash);
